requirements.txt
python-multipart==0.0.6
fastapi==0.95.0
uvicorn==0.21.1
psycopg2-binary==2.9.6
requests==2.28.2  # if you still use 'requests' in llm.py
